requirements.txt
pynput
pydub
pystray
Pillow
thread6
pygame